requirements.txt
msal==1.25.0
requests==2.31.0
pandas==2.1.4
openpyxl==3.1.2
streamlit==1.29.0
python-dateutil==2.8.2
